chore: move controllers to routes/dependencies
