+++ routes/topic/controller/evaluation.py
@@ -1,7 +1,7 @@
 import functools
 import http
 from typing import Sequence, cast
-from controllers.project import ProjectCacheDependency
+from routes.dependencies.project import ProjectCacheDependency
 from modules.api.wrapper import ApiError
 from modules.config.schema.base import SchemaColumnTypeEnum
 from modules.config.schema.schema_variants import TextualSchemaColumn
